Addressing more review comments - split sdf_to_input; overLAID; moved abstrct data sources
