+++ mat_dp_pipeline/data_sources/__init__.py
@@ -1,9 +1,3 @@
-from mat_dp_pipeline.data_sources.definitions import (
-    BaseSource,
-    IndicatorsSource,
-    IntensitiesSource,
-    TargetsSource,
-)
 from mat_dp_pipeline.data_sources.mat_dp_db import (
     MatDPDBIndicatorsSource,
     MatDPDBIntensitiesSource,
