+++ mat_dp_pipeline/data_sources/stored.py
@@ -3,7 +3,7 @@ import shutil
 from pathlib import Path
 
 import mat_dp_pipeline.sdf.standard_data_format as sdf
-from mat_dp_pipeline.data_sources.definitions import (
+from mat_dp_pipeline.abstract_data_sources import (
     IndicatorsSource,
     IntensitiesSource,
     TargetsSource,
