+++ mat_dp_pipeline/data_sources/tmba.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from mat_dp_pipeline.data_sources.definitions import TargetsSource
+from mat_dp_pipeline.abstract_data_sources import TargetsSource
 from mat_dp_pipeline.data_sources.utils import location_to_path
 
 PARAMETER_TO_CATEGORY = {
