+++ mat_dp_pipeline/sdf/__init__.py
@@ -1,2 +1,7 @@
+from mat_dp_pipeline.abstract_data_sources import (
+    IndicatorsSource,
+    IntensitiesSource,
+    TargetsSource,
+)
 from mat_dp_pipeline.sdf.create_sdf import create_sdf
 from mat_dp_pipeline.sdf.standard_data_format import StandardDataFormat, Year
