+++ tests/test_calculation.py
@@ -3,7 +3,6 @@ from io import StringIO
 import pandas as pd
 
 from mat_dp_pipeline.pipeline.calculation import calculate
-from mat_dp_pipeline.pipeline.sdf_to_input import ProcessableInput
 
 
 def test_calculation(calculation_test_input):
